Build serving keys from serving_prepared_statement if not available for TD
